requirements.txt
hugchat
python-dotenv
pyttsx3
SpeechRecognition
pygame
